Change default page from trade to leaderboard
