+++ pages/trade/index.tsx
@@ -1,0 +1,16 @@
+import TradeContainer from "@/containers/trade";
+import Head from "next/head";
+import React from "react";
+
+const LeaderBoard = () => {
+  return (
+    <>
+      <Head>
+        <title>Econia - Trade</title>
+      </Head>
+      <TradeContainer />
+    </>
+  );
+};
+
+export default LeaderBoard;
